Bugfix: Fix Target Information
Previously, the "just_the_tooltip" package received information about the window size based on the parent context, while the calculated coordinates referred to the entire window. This caused the popup displayed at the far right side to appear incorrectly. (In the project where the error was noticed, the popup had X coordinates reduced by the size of the navigation menu on the left side)

List of changes:
- copied functionality that takes the size of the browser window from material "Tooltip" widget (from "_createNewEntry" method) and placed it inside "_getTargetInformation" method (just_the_tooltip.dart)